Moved ResourceTracker to core
Removed duplicate astral power display.
Changed astral power to show per minute instead of just total.
