+++ src/Parser/Core/Modules/ResourceTracker/ResourceSuggest.js
