+++ src/Parser/Core/Modules/ResourceTracker/ResourceTracker.js
