+++ src/Parser/Rogue/Subtlety/Modules/RogueCore/ComboPointTracker.js
@@ -1,7 +1,7 @@
 import RESOURCE_TYPES from 'common/RESOURCE_TYPES';
 import Combatants from 'Parser/Core/Modules/Combatants';
 
-import ResourceTracker from '../ResourceTracker/ResourceTracker';
+import ResourceTracker from 'Parser/Core/Modules/ResourceTracker/ResourceTracker';
 
 class ComboPointTracker extends ResourceTracker {
   static dependencies = {
