+++ src/Parser/Rogue/Subtlety/Modules/RogueCore/EnergyDetails.js
@@ -10,8 +10,8 @@ import Wrapper from 'common/Wrapper';
 
 import EnergyTracker from './EnergyTracker';
 
-import ResourceBreakdown from '../ResourceTracker/ResourceBreakdown';
-import resourceSuggest from '../ResourceTracker/ResourceSuggest';
+import ResourceBreakdown from 'Parser/Core/Modules/ResourceTracker/ResourceBreakdown';
+import resourceSuggest from 'Parser/Core/Modules/ResourceTracker/ResourceSuggest';
 
 class EnergyDetails extends Analyzer {
   static dependencies = {
